stream: move inline tests
